Fixes for newer android
Newer android does not have init.d and requires a different
style of rc script.

On android9, this creates /system/etc/init.d so that the 55atlas can be
placed there... and then creates a 55atlas.rc and places it in
/system/etc/init which will run the /system/etc/init.d much like it
does on android7.

Also, / needs remounted, not /system, etc.